Saved my progress
I had actually planned on adding a remote to my previous Git repo, but when I saw GitHub had a way to include a license from the get-go, I figured, what the puck.  Now the 6502 simulator I used is MIT licensed (hopefully that is not a problem) but *my* code is free for everyone, no strings attached.  Anyway, here's where I left off:

* The memory map is actually close to being finalized.  I have a few other things I might want to add, but they're minor.
* The screen is working, automatically updating based on what's in RAM.
* The keyboard needs a bit of work; I need a function to map ASCII to PETSCII, cuz right now if you type an A you get character #65, which is not an A (A is #1 in PETSCII).
* I also want to try and add mouse support, cuz the mouse is a useful tool (even though I prefer the keyboard) and most people who have one have the other.  And think of games like Breakout, Pong, Space Invaders etc. - in the terminal, those make a lot more sense to be mouse-driven (or at least mouse-supported).
* And really, it's about time to start trying some 6502 code!