Oh puck yeah!
I moved the logic for reading the keyboard to OnRead, and all of a sudden, the VM is *ridiculously* fast!  I almost feel like I'm just in the terminal; as in no delay at all!  And here I thought the performance issues were with updateDisplay!  Well, thank you Lord! :)

Oh yeah, and I've also added the check for max rows/columns, and so far it looks right, even in cool-retro term!  That is so cool! :)