Fixed the bug :)
Since the terminal's ability to monitor keyboard input is limited to key presses, and only printable characters, I've set it up so the Geek-Rig stores the LAST key pressed, not the CURRENT key pressed.
This seems to work perfectly, cuz I can now do LDA #0 and STA KEY if I don't want endless "key repeat" (which is the case in my Space Invaders game - not as important in Viper, and now that can be different for each game).

Oh yeah, and my player can move and shoot :)