Wrote a little program I want to test on the CRT.
It prints the columns at the top left, and the on the row below it.
Works great (apart from a PRINT subroutine I haven't cracked yet) on
both xterm and cool-retro-term.  I'm still learning the ins and outs
of 16-bit mathing on 8-bit registers, very very hard stuff for anyone
who (like me) is not a super algebraniac.  Obviously, my skills are
much more language-related lol... anyway, the point of crawling these
mathing mazes is so that I can create games that work at multiple
screen dimensions; for example, my cool-retro-term is at 40x11, which
is great for me (for my vision) but probably not for most gamers.  On
my CRT, I expect it'll be something like 40x15, but most people will
probably run this at full size, 80x24 (but keep in mind the 40x24 max
the 4000 uses for RAM reasons). Anyway, the interesting thing here is
how these size limits affect games.  Like look at Space Invaders: the
enemy space ships have a lot less distance to travel to hit the bottom
on a 40x11 screen vs. 40x22.  Same deal with Viper (though that game
doesn't look like it checks for hitting the border).